Moved pyramid fixture into its own package, pytest-pyramid
