+++ pytest-pyramid/README.rst
@@ -1,0 +1,8 @@
+Py.test Pyramid Server Fixture
+==============================
+
+                      
+Documentation
+=============
+
+Coming soon!
