+++ pytest-pyramid/setup.cfg
@@ -1,0 +1,18 @@
+[egg_info]
+# This adds the 'dev' marker to versions.
+# This must conform to PEP-386:
+# http://www.python.org/dev/peps/pep-0386/#the-new-versioning-algorithm
+tag_build = .dev1
+
+[upload_docs]
+upload-dir = build/sphinx/html
+show-response = 1
+
+[pytest]
+# This section sets configuration for all invocations of py.test,
+# both standalone cmdline and running via setup.py
+norecursedirs = 
+	.git 
+	*.egg 
+	build 
+	dist
